src/agents/crewai/04-financial-research/requirements.txt
crewai[tools]>=0.100.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.0
pydantic>=2.0.0
